Integrate smart when lazy fixes to tuple2
